+++ src/components/weather-test.html
@@ -16,7 +16,7 @@
     </div>
 
     <script type="module" src="../ts/weather.ts"></script>
-    <script type="module" src="../ts/shortWeatherApi.ts"></script>
-    <script type="module" src="../ts/weatherApi.ts"></script>
+    <script type="module" src="../ts/api/shortWeatherApi.ts"></script>
+    <script type="module" src="../ts/api/weatherApi.ts"></script>
   </body>
 </html>
